Layout and Animate Hierarchical Relationships
Adds hierarchical layouts and animations for all relationships. Most
recently added view is given highest focus by giving it a 2/3 display on
the right. The second most recent view is given 1/3 on the left. If only
one view exists it gets full screen. Animations occur from right to left
as views are added.

Change-Id: I129e03e47cc1448a27cd1ac6e8d0436bf6df96ed