Simplify SimulatedPositioned Constructors
Change SimulatedPositioned constructor to take Offset & Size or a Rect.
Add initRect to SimulatedPositioned to give it a starting location for
its animation that is different than its target, initialize with
animation rather than having to place it on the stack, then queue up an
immediate change to move it. Also introduces a convenience function for
generating SurfaceWidgets in the SurfaceLayout.

Change-Id: I1987ea5b63ee9891fb3ceb43419e34bc337d4b8b