[maxwell] Suggestion Engine refactor
We are refactoring the suggestion engine and the ProposalPublisher
such that registering an AskHandler will not also require consumers
to register a ProposalPublisher.  This CL alters consumers of those
methods to use the new pattern.  For User Shell, this updates the
services passed through to remove ProposalPublisher and include
IntelligenceServices instead.  IntelligenceServices can be used to
acquire a ProposalPublisher when needed.

MW-233 #done

Change-Id: Iecc3fefffe408b5ca00c117c92e8f98a5aeb743e