[maxwell] Suggestion Engine refactor
We are refactoring the suggestion engine and the ProposalPublisher
such that registering an AskHandler will not also require consumers
to register a ProposalPublisher.  This CL alters consumers of those
methods to use the new pattern.  For User Shell, this updates the
services passed through to remove ProposalPublisher and include
IntelligenceServices instead.  IntelligenceServices can be used to
acquire a ProposalPublisher when needed.

Peridot works now, so this can go in.

MW-233 #done

This reverts commit 87119c62ead849752d8a33a20aa8c12eeabca934.

Change-Id: I0f46806f8fb35a7f762483b91c31ebc1669b7fc0